requirements.txt
APScheduler==3.6.3
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.1
decorator==5.1.1
hiredis==2.2.2
idna==3.4
pycparser==2.21
python-dotenv==1.0.0
python-telegram-bot==13.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.1
requests==2.28.2
six==1.16.0
tornado==6.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.15
vk-api==11.9.9
